Move tv cog components and models to their modules
