+++ flanders/components/__init__.py
@@ -1,0 +1,1 @@
+"""All v2 components used by FlandersBOT."""
