+++ flanders/components/tv.py
@@ -1,0 +1,239 @@
+from abc import abstractmethod
+from typing import Sequence
+
+import compuglobal
+import discord
+
+from flanders.cogs.events import Events
+from flanders.models import TVReferenceState
+
+
+class TVContentView(discord.ui.LayoutView):
+    def __init__(self, content_url: str, episode_title: str, episode_url: str, episode_details: str, author: str):
+        super().__init__()
+        self.content_url = content_url
+
+        container = discord.ui.Container()
+
+        # Episode title (Simpsons Roasting on an Open Fire)
+        title = discord.ui.TextDisplay(content=f"### **[{episode_title}]({episode_url})**")
+        container.add_item(title)
+
+        # Episode details (S01E01 - 00:00)
+        details = discord.ui.TextDisplay(f"-# {episode_details}")
+        container.add_item(details)
+
+        # Content gif/comic
+        tv_content = discord.ui.MediaGallery()
+        tv_content.add_item(media=self.content_url)
+        container.add_item(tv_content)
+
+        # Generated by @username
+        footer = discord.ui.TextDisplay(content=f"-# *Generated by {author}*")
+        container.add_item(footer)
+
+        self.add_item(container)
+
+
+class SearchResult(discord.SelectOption):
+    def __init__(self, frame: compuglobal.Frame, index: int, api_title: str, state: TVReferenceState):
+        summary = state.api_cache.get(frame.key)
+        title = summary.title if summary is not None else "Unknown Title"
+        super().__init__(label=f"{index}. {title}", description=f"{frame.key} - {frame.get_real_timestamp()}")
+
+
+class GifBuilderView(discord.ui.LayoutView):
+    def __init__(self, options: Sequence[discord.SelectOption], state: TVReferenceState, image_url: str):
+        super().__init__()
+        self.state = state
+
+        # Add the search results dropdown
+        action_row = discord.ui.ActionRow()
+        action_row.add_item(SearchResultDropdown(options, state))
+        self.add_item(action_row)
+
+        self.gallery = discord.ui.MediaGallery()
+        self.gallery.add_item(media=image_url)
+        self.add_item(self.gallery)
+
+        # Add the customisation and post content buttons
+        action_row = discord.ui.ActionRow()
+        action_row.add_item(GenerateGifButton(state))
+        action_row.add_item(GenerateComicButton(state))
+        action_row.add_item(CustomiseCaptionButton(state))
+        self.add_item(action_row)
+
+    def update_image(self, image_url: str):
+        self.gallery.clear_items()
+        self.gallery.add_item(media=image_url)
+
+
+class SearchResultDropdown(discord.ui.Select):
+    def __init__(self, options: Sequence[discord.SelectOption], state: TVReferenceState):
+        self.state = state
+        super().__init__(placeholder="Choose the best match...", min_values=1, max_values=1, options=list(options))
+
+    async def callback(self, interaction: discord.Interaction):
+        await interaction.response.defer(ephemeral=True)
+        selected_index = 0
+        for i in range(len(self.options)):
+            if self.options[i].value == self.values[0]:
+                selected_index = i
+                self.options[i].default = True
+            else:
+                self.options[i].default = False
+
+        # Update selected index state
+        self.state.set_index(selected_index)
+        if self.view is not None:
+            self.view.update_image(await self.state.get_comic_strip_url())
+            await interaction.edit_original_response(view=self.view)
+
+
+class GenerateButton(discord.ui.Button):
+    def __init__(self, label: str, style: discord.ButtonStyle, state: TVReferenceState):
+        self.state = state
+        super().__init__(label=label, style=style)
+
+    async def callback(self, interaction: discord.Interaction):
+        await interaction.response.defer(ephemeral=True)
+
+        # Disable comic/gif builder view elements
+        if self.view is not None:
+            for child in self.view.walk_children():
+                if isinstance(child, (discord.ui.Button, discord.ui.Select)):
+                    child.disabled = True
+
+            await interaction.edit_original_response(view=self.view)
+
+        # Generate gif
+        screencap = await self.state.get_screencap()
+        emoji = await Events.use_emoji(interaction, "<a:loading:410316176510418955>", "⌛")
+
+        interaction_channel = interaction.channel
+
+        original = None
+        if isinstance(interaction_channel, discord.TextChannel):
+            original = await interaction_channel.send(f"Generating {screencap.frame.key}... {emoji}")
+
+        await self.state.cache_screencap()
+
+        content_url = await self.get_content_url()
+        content_view = TVContentView(
+            content_url=content_url,
+            episode_title=screencap.episode.title,
+            episode_url=f"{self.state.api.BASE_URL}/episode/{screencap.frame.key}/{screencap.frame.timestamp}",
+            episode_details=f"{screencap.frame.key} - {screencap.get_real_timestamp()}",
+            author=interaction.user.mention,
+        )
+
+        if original is not None:
+            await original.edit(content=None, view=content_view, allowed_mentions=discord.AllowedMentions.none())
+
+    @abstractmethod
+    async def get_content_url(self) -> str:
+        pass
+
+
+class GenerateComicButton(GenerateButton):
+    def __init__(self, state: TVReferenceState):
+        super().__init__(label="Send Comic", style=discord.ButtonStyle.secondary, state=state)
+
+    async def get_content_url(self):
+        return await self.state.get_comic_strip_url()
+
+
+class GenerateGifButton(GenerateButton):
+    def __init__(self, state: TVReferenceState):
+        super().__init__(label="Send Gif", style=discord.ButtonStyle.primary, state=state)
+
+    async def get_content_url(self):
+        return await self.state.get_gif_url()
+
+
+class CustomiseCaptionButton(GenerateButton):
+    def __init__(self, state: TVReferenceState):
+        self.state = state
+        super().__init__(state=state, label="Edit Captions", style=discord.ButtonStyle.secondary)
+
+    async def callback(self, interaction: discord.Interaction):
+        if self.view is not None:
+            subtitles = await self.state.get_subtitles()
+            duration = await self.state.get_total_duration()
+            modal = CustomiseCaptionModal(
+                total_duration=duration, state=self.state, view=self.view, subtitles=subtitles
+            )
+
+            try:
+                await interaction.response.send_modal(modal)
+
+            except Exception as e:
+                print(e)
+
+
+class CustomiseCaptionModal(discord.ui.Modal, title="Customise caption:"):
+    def __init__(
+        self, total_duration: int, state: TVReferenceState, view: GifBuilderView, subtitles: list[compuglobal.Subtitle]
+    ):
+        super().__init__()
+        self.state = state
+        self.view = view
+
+        labels = ["first", "second", "third", "fourth", "fifth"]
+
+        for i, subtitle in enumerate(subtitles):
+            duration = subtitle.end_timestamp - subtitle.start_timestamp
+            seconds = duration / 1000
+            custom_caption = discord.ui.TextInput(
+                label=f"Customise {labels[i]} caption ({seconds:.1f} sec):",
+                default=subtitle.content,
+                style=discord.TextStyle.short,
+                placeholder="Enter your caption...",
+                required=False,
+                max_length=150,
+            )
+            self.add_item(custom_caption)
+
+        self.merge_caption_checkbox = discord.ui.Checkbox(default=False)
+        merge_caption = discord.ui.Label(
+            text=f"Combine above captions ({total_duration / 1000:.1f} sec)",
+            component=self.merge_caption_checkbox,
+        )
+        self.add_item(merge_caption)
+
+    def get_captions(self):
+        for child in self.children:
+            if isinstance(child, discord.ui.TextInput):
+                yield child.value
+
+    async def get_subtitles(self) -> list[compuglobal.Subtitle]:
+        screencap = await self.state.get_screencap()
+        return [
+            subtitle.model_copy(update={"content": new_content})
+            for subtitle, new_content in zip(screencap.subtitles, self.get_captions(), strict=True)
+        ]
+
+    async def get_merged_subtitles(self) -> list[compuglobal.Subtitle]:
+        subtitles = await self.get_subtitles()
+
+        content = " ".join(subtitle.content for subtitle in subtitles)
+        start_timestamp = min(subtitle.start_timestamp for subtitle in subtitles)
+        end_timestamp = max(subtitle.end_timestamp for subtitle in subtitles)
+
+        return [
+            subtitle.model_copy(
+                update={"content": content, "start_timestamp": start_timestamp, "end_timestamp": end_timestamp}
+            )
+            for subtitle in subtitles
+        ]
+
+    async def on_submit(self, interaction: discord.Interaction):
+        await interaction.response.defer(ephemeral=True)
+
+        self.state.custom_subtitles = (
+            await self.get_merged_subtitles() if self.merge_caption_checkbox.value else await self.get_subtitles()
+        )
+
+        # Update image/comic preview
+        self.view.update_image(await self.state.get_comic_strip_url())
+        await interaction.edit_original_response(view=self.view)
