+++ flanders/models/__init__.py
@@ -1,3 +1,4 @@
+from flanders.models.reference import TVReferenceState
 from flanders.models.trivia_category import FuturamaTrivia, RickAndMortyTrivia, SimpsonsTrivia, TriviaCategory
 
-__all__ = ["FuturamaTrivia", "RickAndMortyTrivia", "SimpsonsTrivia", "TriviaCategory"]
+__all__ = ["FuturamaTrivia", "RickAndMortyTrivia", "SimpsonsTrivia", "TriviaCategory", "TVReferenceState"]
